+++ analysis/tools/InspectionTool/src/test/kotlin/de/maibornwolff/codecharta/tools/inspection/PipedInputTestHelper.kt
@@ -1,0 +1,29 @@
+package de.maibornwolff.codecharta.tools.inspection
+
+import de.maibornwolff.codecharta.tools.inspection.InspectionTool.Companion.mainWithInOut
+import java.io.ByteArrayInputStream
+import java.io.ByteArrayOutputStream
+import java.io.InputStream
+import java.io.PrintStream
+
+fun executeForOutput(input: String, args: Array<String> = emptyArray()) = outputAsString(input) { inputStream, outputStream ->
+    mainWithInOut(inputStream, outputStream, args)
+}
+
+fun outputAsString(input: String, aMethod: (input: InputStream, output: PrintStream) -> Unit) = outputAsString(
+    ByteArrayInputStream(input.toByteArray()),
+    aMethod
+)
+
+fun outputAsString(
+    inputStream: InputStream = System.`in`,
+    aMethod: (
+        input: InputStream,
+        output: PrintStream
+    ) -> Unit
+) = ByteArrayOutputStream().use { baOutputStream ->
+    PrintStream(baOutputStream).use { outputStream ->
+        aMethod(inputStream, outputStream)
+    }
+    baOutputStream.toString()
+}